parser/tests/tree_construction/basic.dat

#data
<!doctype html><input type="hidden"><frameset>
#errors
(1,46): unexpected-start-tag
(1,46): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
